requirements.txt
streamlit==1.40.0
python-dotenv==1.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
openai==1.52.0
tiktoken==0.8.0
pypdf==5.0.0
pinecone-client==5.0.1
tqdm==4.66.0